Bank, Enterprise, WalletType updated. ORM modified

+++ src/controllers/user.js
@@ -3,32 +3,34 @@ const uuid = require('uuid');
 const models = require("../../models");
 
 const createUser = async (req, res) => {
-  //const name = req.body.name;
-  //const surname = req.body.surname;
-  //const email = req.body.email;
-  //const username = req.body.username;
   try {
-    const { name, surname, email, username } = req.body;
+    const { name, surname, email } = req.body;
+    const username = helpers.hasNoSpaces(req.body.username);
     const password = await helpers.encryptPassword(req.body.password);
-    const user = await models.User.create({
-      Usr_name: name,
-      Usr_surname: surname,
-      Usr_email: email,
-      Usr_username: username,
-      Usr_password: password
-    });
 
-    const wallet = await models.Wallet.create({
-      Wal_id: uuid.v4(),
-      Usr_id: user.Usr_id,
-      Wtyp_id: 1,
-      Wal_balance: 0.00,
-      Wal_state: "Active"
-    });
+    if (username) {
+      const user = await models.User.create({
+        Usr_name: name,
+        Usr_surname: surname,
+        Usr_email: email,
+        Usr_username: username,
+        Usr_password: password
+      });
 
-    return res.status(201).json({
-      user: user,
-    });
+      const wallet = await models.Wallet.create({
+        Wal_id: uuid.v4(),
+        Usr_id: user.Usr_id,
+        Wtyp_id: 1,
+        Wal_balance: 0.00,
+        Wal_state: "Active"
+      });
+
+      return res.status(201).json({
+        user: user,
+        wallet: wallet
+      });
+    }
+    return res.status(400).send("Username can't contain spaces");
   } catch (error) {
     return res.status(500).json({ error: error.message })
   }
@@ -61,26 +63,27 @@ const getUserByUsername = async (req, res) => {
 const updateUser = async (req, res) => {
   try {
     const { username } = req.params;
-    const name = req.body.name;
-    const surname = req.body.surname;
-    const usrname = req.body.username;
-    const email = req.body.email;
+    const usrname = helpers.hasNoSpaces(req.body.username);
+    const { name, surname, email } = req.body;
     const password = await helpers.encryptPassword(req.body.password);
 
-    const [updated] = await models.User.update({
-      Usr_name: name,
-      Usr_surname: surname,
-      Usr_email: email,
-      Usr_username: usrname,
-      Usr_password: password
-    }, {
-      where: { Usr_username: username }
-    });
-    if (updated) {
-      const updatedUser = await models.User.findOne({ where: { Usr_username: usrname } });
-      return res.status(200).json({ user: updatedUser });
+    if (usrname) {
+      const [updated] = await models.User.update({
+        Usr_name: name,
+        Usr_surname: surname,
+        Usr_email: email,
+        Usr_username: usrname,
+        Usr_password: password
+      }, {
+        where: { Usr_username: username }
+      });
+      if (updated) {
+        const updatedUser = await models.User.findOne({ where: { Usr_username: usrname } });
+        return res.status(200).json({ user: updatedUser });
+      }
+      throw new Error('User not updated');
     }
-    throw new Error('User not found');
+    return res.status(400).send("Username can't contain spaces");
   } catch (error) {
     return res.status(500).send(error.message);
   }
@@ -90,26 +93,24 @@ const deleteUser = async (req, res) => {
   try {
     const { username } = req.params;
     const findUser = await models.User.findOne({ where: { Usr_username: username } });
-    const wallets = await models.Wallet.findAll({ where: { Usr_id: findUser.Usr_id } });
 
-    const deleted = await models.User.destroy({
-      where: { Usr_username: username }
-    });
-    if (deleted) {
-      wallets.forEach(async (element) => {
-        var del = await models.Wallet.destroy({
-          where: { Wal_id: element.Wal_id }
-        });
-        if (!del) {
-          throw new Error("Associated wallet "+element.Wal_id+" not deleted");
-        }
+    if (findUser) {
+      const wallets = await models.Wallet.findAll({ where: { Usr_id: findUser.Usr_id } });
+      
+      if (wallets.length > 0) {
+        return res.status(400).send("This User has associated Wallets. Please delete them first");
+      }
+      const deleted = await models.User.destroy({
+        where: { Usr_username: username }
       });
-      return res.status(200).send("User "+username+" deleted");
+      if (deleted) {
+        return res.status(200).send("User " + username + " deleted");
+      }
     }
-    throw new Error("User not found");
-  } catch (error) {
-    return res.status(500).send(error.message);
-  }
+    return res.status(404).send("Specified User not found");
+} catch (error) {
+  return res.status(500).send(error.message);
+}
 };
 
 const validateUser = async (req, res) => {
@@ -125,6 +126,10 @@ const validateUser = async (req, res) => {
             {
               model: models.Movement,
               as: 'modifies sender',
+            },
+            {
+              model: models.Movement,
+              as: 'modifies recipient', 
             }
           ]
         }
