+++ src/controllers/wallet.js
@@ -66,6 +66,40 @@ const getAllWallets = async (req, res) => {
     }
 }
 
+const updateWallet = async (req, res) => {
+    try {
+        const { wal_id } = req.params;
+        const { wallettype, balance, NIT } = req.body;
+        const state = helpers.isWalletState(req.body.state); //Limited in the future with new model creation
+        const findWtyp = await models.WalletType.findOne({ where: { Wtyp_id: wallettype } });
+
+        if (state == null) {
+            return res.status(400).send('Not a valid wallet state (Active, Inactive)');
+        } else if (!findWtyp) {
+            return res.status(400).send('Not a valid wallet type');
+        }
+
+        const [updated] = await models.Wallet.update({
+            Wtyp_id: wallettype,
+            Ent_NIT: NIT,
+            Wal_balance: balance,
+            Wal_state: state
+        }, {
+            where: { Wal_id: wal_id }
+        });
+
+        const updatedWallet = await models.Wallet.findOne({ where: { Wal_id: wal_id } });
+        if (updated) {
+            return res.status(200).json({ wallet: updatedWallet });
+        } else if (updatedWallet) {
+            return res.status(400).send("Not updated: Update data is the same");
+        }
+        return res.status(404).send("Wallet not found");
+    } catch (error) {
+        return res.status(500).send(error.message);
+    }
+};
+
 const updateWalletState = async (req, res) => {
     try {
         const { username, wal_id } = req.params;
@@ -118,6 +152,12 @@ const deleteWallet = async (req, res) => {
 
         if (findUser) {
             if (findWallet) {
+                const movement_sender = await models.Movement.findAll({ where: { Wal_id_sender: wal_id } });
+                const movement_recipient = await models.Movement.findAll({ where: { Wal_id_recipient: wal_id } });
+
+                if (movement_sender.length > 0 || movement_recipient.length > 0 ) {
+                    return res.status(400).send("This Wallet has associated Movements. Please delete them first");
+                }
                 const val = await helpers.matchPassword(password, findUser.Usr_password);
                 if (val) {
                     const deleted = await models.Wallet.destroy({
@@ -142,6 +182,7 @@ const deleteWallet = async (req, res) => {
 module.exports = {
     createWallet,
     getAllWallets,
+    updateWallet,
     updateWalletState,
     getAllWalletsByUsername,
     deleteWallet
